Move get_local_track_handles and local_stream_update_result to InnerRoom
